+++ src/aging/aging.py
@@ -5,7 +5,13 @@
 from argparse import Namespace
 
 ##############################################################################
+# Pyperclip imports.
+from pyperclip import PyperclipException
+from pyperclip import copy as copy_to_clipboard
+
+##############################################################################
 # Textual imports.
+from textual import on
 from textual.app import InvalidThemeError, ScreenStackError
 
 ##############################################################################
@@ -19,6 +25,7 @@ from .data import (
     load_configuration,
     update_configuration,
 )
+from .messages import CopyToClipboard
 from .screens import Main
 
 
@@ -80,6 +87,27 @@ class AgiNG(EnhancedApp[None]):
     def get_default_screen(self) -> Main:
         """Get the main screen for the application."""
         return Main(self._arguments)
+
+    @on(CopyToClipboard)
+    def _copy_text_to_clipbaord(self, message: CopyToClipboard) -> None:
+        """Copy some text into the clipboard.
+
+        Args:
+            message: The message requesting the text be copied.
+        """
+        # First off, use Textual's own copy to clipboard facility. Generally
+        # this will work in most terminals, and if it does it'll likely work
+        # best, getting the text through remote connections to the user's
+        # own environment.
+        self.copy_to_clipboard(message.text)
+        # However, as a backup, use pyerclip too. If the above did fail due
+        # to the terminal not supporting the operation, this might work.
+        try:
+            copy_to_clipboard(message.text)
+        except PyperclipException:
+            pass
+        # Give the user some feedback.
+        self.notify(f"Copied {message.description or ''}".strip())
 
 
 ### aging.py ends here